fix/B2BCHCKOUT-26_app-review
