+++ graphql/directives.graphql
@@ -1,0 +1,1 @@
+directive @withAdminAccess on FIELD | FIELD_DEFINITION
