+++ manifest.json
@@ -25,6 +25,13 @@
   "credentialType": "absolute",
   "policies": [
     {
+      "name": "outbound-access",
+      "attrs": {
+        "host": "portal.vtexcommercestable.com.br",
+        "path": "/api/*"
+      }
+    },
+    {
       "name": "vtex.graphql-server:resolve-graphql"
     },
     {
