+++ node/resolvers/constants.ts
@@ -1,0 +1,7 @@
+export const VBASE_BUCKET = 'b2b-checkout-settings'
+export const VBASE_SETTINGS_FILE = 'settings'
+export const DEFAULTS = {
+  showPONumber: false,
+  hasPONumber: false,
+  showQuoteButton: false,
+}
