+++ node/resolvers/directives.ts
@@ -1,0 +1,5 @@
+import { WithPermissions } from './directives/withAdminAccess'
+
+export const schemaDirectives = {
+  withAdminAccess: WithPermissions as any,
+}
